backend/python-backend/requirements.txt
beautifulsoup4==4.13.3
Flask==3.1.0
Flask_Cors==5.0.0
groq==0.18.0
joblib==1.4.2
numpy==2.2.2
openai==1.61.1
pandas==2.2.3
pydantic==2.10.6
python-dotenv==1.0.1
Requests==2.32.3
tensorflow==2.18.0
